# backend/docs_design/openapi.yaml
...
openapi: 3.1.0
info:
  title: BoMesh API
  version: 1.0.0
  description: Contract-first target for workspace knowledge, IAM, integrations, and governed ingestion.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/connections:
    get:
      tags: [connections]
      operationId: listConnections
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Search'}
      responses:
        '200':
          description: Connections
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConnectionPage'}
    post:
      tags: [connections]
      operationId: createConnection
      x-authorization-rule: Personal connections belong to the caller; workspace-owned connections require source.manage.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConnectionCreate'}
      responses:
        '201':
          description: Created Connection
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Connection'}
  /api/v1/connections/authorizations:
    post:
      tags: [connections]
      operationId: createConnectionAuthorization
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConnectionAuthorizationCreate'}
      responses:
        '201':
          description: Provider authorization target
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConnectionAuthorization'}
  /api/v1/connections/{connection_id}:
    parameters:
      - {$ref: '#/components/parameters/ConnectionId'}
    get:
      tags: [connections]
      operationId: getConnection
      responses:
        '200':
          description: Connection detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Connection'}
    patch:
      tags: [connections]
      operationId: updateConnection
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ConnectionUpdate'}
      responses:
        '200':
          description: Updated Connection
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Connection'}
    delete:
      tags: [connections]
      operationId: deleteConnection
      responses:
        '204': {description: Connection tombstoned}
  /api/v1/connections/{connection_id}/validate:
    post:
      tags: [connections]
      operationId: validateConnection
      parameters:
        - {$ref: '#/components/parameters/ConnectionId'}
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ConnectionValidation'}
  /api/v1/connections/{connection_id}/resources:
    get:
      tags: [connections]
      operationId: listConnectionResources
      parameters:
        - {$ref: '#/components/parameters/ConnectionId'}
        - name: parent_id
          in: query
          schema: {type: string}
        - {$ref: '#/components/parameters/Search'}
      responses:
        '200':
          description: Provider resources
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProviderResourcePage'}
  /api/v1/connections/{connection_id}/sources:
    post:
      tags: [sources]
      operationId: createConnectionSource
      parameters:
        - {$ref: '#/components/parameters/ConnectionId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SourceCreate'}
      responses:
        '201':
          description: Created Source
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Source'}
...
  /api/v1/approval-requests:
    get:
      tags: [governance]
      operationId: listApprovalRequests
      x-authorization-rule: 'Own requests, or reviewable types: access.manage for resource_access, source.manage for plugin_installation.'
      responses:
        '200': {description: Approval Requests, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalRequestPage'}}}}
    post:
      tags: [governance]
      operationId: createApprovalRequest
      x-authorization-rule: Own resource_access request; plugin_installation creation requires source.manage.
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalRequestCreate'}}}
      responses:
        '201': {description: Created Approval Request, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalRequest'}}}}
  /api/v1/approval-requests/{approval_request_id}:
    parameters:
      - {$ref: '#/components/parameters/ApprovalRequestId'}
    get:
      tags: [governance]
      operationId: getApprovalRequest
      x-authorization-rule: 'Own request, or reviewer permission: access.manage for resource_access, source.manage for plugin_installation.'
      responses:
        '200': {description: Approval Request, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalRequest'}}}}
    patch:
      tags: [governance]
      operationId: updateApprovalRequest
      x-authorization-rule: Pending only. Requester may cancel; approve/deny requires access.manage for resource_access or source.manage for plugin_installation. Reviewers may also cancel.
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalRequestUpdate'}}}
      responses:
        '200': {description: Updated Approval Request, content: {application/json: {schema: {$ref: '#/components/schemas/ApprovalRequest'}}}}
...
  /api/v1/access-sessions:
    get:
      tags: [governance]
      operationId: listAccessSessions
      x-required-permissions: [audit.read]
      description: >-
        Sign-in sessions of the caller's active workspace (kind user only),
        newest first (`created_at` desc, then `id`).
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - name: status
          in: query
          description: '`active`: effectively active now; `ended`: every other session.'
          schema: {type: string, enum: [active, ended]}
        - name: user_id
          in: query
          schema: {type: string, format: uuid}
        - name: search
          in: query
          description: Case-insensitive match on the user's email or display name.
          schema: {type: string}
      responses:
        '200': {description: Access sessions, content: {application/json: {schema: {$ref: '#/components/schemas/AccessSessionPage'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
    Search: {name: search, in: query, schema: {type: string, maxLength: 255}}
...
    ConnectionId: {name: connection_id, in: path, required: true, schema: {type: string, format: uuid}}
...
    ApprovalRequestId: {name: approval_request_id, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
...
    Forbidden:
      description: Insufficient permission
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    ValidationError:
      description: Request validation failed
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id, details]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string, format: uuid}
        details: {type: object, additionalProperties: true}
    PageFields:
      type: object
      required: [page, page_size, total]
      properties:
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
...
    Connection:
      type: object
      required: [id, connector_key, display_name, owner_type, status, source_count]
      properties:
        id: {type: string, format: uuid}
        connector_key: {type: string}
        display_name: {type: string}
        owner_type: {type: string, enum: [user, workspace]}
        owner_user_id: {type: [string, 'null'], format: uuid}
        status: {type: string, enum: [draft, connected, expired, reauth_required, revoked, error, disconnected]}
        source_count: {type: integer, minimum: 0}
        config: {type: object, additionalProperties: true}
        account:
          type: object
          description: Whose account this is, never its secret or provider identifier.
          required: [label, resource_label]
          properties:
            label: {type: [string, 'null']}
            resource_label: {type: [string, 'null']}
        browsable:
          type: boolean
          description: >-
            `GET /connections/{connection_id}/resources` can list what this
            connection reaches: through a provider sign-in, or through a
            credential the connector can browse with (a Confluence API token).
        status_detail: {type: [string, 'null']}
        connected_at: {type: [string, 'null'], format: date-time}
        last_checked_at: {type: [string, 'null'], format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    ConnectionCreate:
      type: object
      required: [connector_key, display_name]
      properties:
        connector_key: {type: string}
        display_name: {type: string}
        owner_type: {type: string, enum: [user, workspace], default: workspace}
        config: {type: object, additionalProperties: true}
        credentials: {type: object, writeOnly: true, additionalProperties: true}
    ConnectionUpdate:
      type: object
      minProperties: 1
      properties:
        display_name: {type: string}
        status: {type: string, enum: [disconnected]}
        config: {type: object, additionalProperties: true}
        credentials: {type: object, writeOnly: true, additionalProperties: true}
    ConnectionPage:
      allOf:
        - {$ref: '#/components/schemas/PageFields'}
        - type: object
          required: [items]
          properties: {items: {type: array, items: {$ref: '#/components/schemas/Connection'}}}
    ConnectionAuthorizationCreate:
      type: object
      required: [connector_key, owner_type]
      properties:
        connector_key: {type: string}
        owner_type: {type: string, enum: [user, workspace]}
        connection_id: {type: [string, 'null'], format: uuid}
    ConnectionAuthorization:
      type: object
      required: [authorization_url, nonce]
      properties:
        authorization_url: {type: string, format: uri}
        nonce: {type: string}
    ConnectionValidation:
      type: object
      required: [valid, status]
      properties:
        valid: {type: boolean}
        status: {type: string}
...
    ProviderResourcePage:
      type: object
      required: [items]
      properties: {items: {type: array, items: {type: object, additionalProperties: true}}}
    Source:
      type: object
      required: [id, connection_id, collection_id, sync_mode, status]
      properties:
        id: {type: string, format: uuid}
        connection_id: {type: string, format: uuid}
        collection_id: {type: string, format: uuid}
        display_name: {type: [string, 'null']}
        resource_type: {type: [string, 'null']}
        external_resource_id: {type: [string, 'null']}
        sync_mode: {type: string, enum: [manual, scheduled]}
        status: {type: string, enum: [ready, paused, failed, connection_required, disabled]}
        schedule: {anyOf: [{$ref: '#/components/schemas/Schedule'}, {type: 'null'}]}
        sync:
          description: The latest sync (it registers, updates and removes Documents; it never processes). Null before the first sync.
          anyOf:
            - type: object
              required: [status]
              properties:
                status: {type: string, enum: [running, succeeded, failed]}
                last_synced_at: {type: [string, 'null'], format: date-time}
                error: {type: [string, 'null'], description: User-safe reason the sync failed}
                added: {type: integer, minimum: 0}
                updated: {type: integer, minimum: 0}
                removed: {type: integer, minimum: 0}
                failed: {type: integer, minimum: 0}
            - {type: 'null'}
        pending_documents: {type: integer, minimum: 0, description: The Source's Documents that are pending or outdated}
    SourceCreate:
      type: object
      required: [collection_id]
      properties:
        collection_id: {type: string, format: uuid}
        display_name: {type: [string, 'null']}
        resource_type: {type: [string, 'null']}
        external_resource_id: {type: [string, 'null']}
        config: {type: object, additionalProperties: true}
        schedule: {anyOf: [{$ref: '#/components/schemas/SchedulePut'}, {type: 'null'}]}
...
    SchedulePut:
      type: object
      required: [schedule_type, cron_expression, enabled, overlap_policy]
      properties:
        schedule_type: {type: string, enum: [cron, interval]}
        cron_expression: {type: string}
        timezone: {type: [string, 'null']}
        enabled: {type: boolean}
        overlap_policy: {type: string, enum: [skip, queue, replace]}
...
    Schedule:
      allOf:
        - {$ref: '#/components/schemas/SchedulePut'}
        - type: object
          required: [id]
          properties:
            id: {type: string, format: uuid}
            next_run_at: {type: [string, 'null'], format: date-time}
            last_run_at: {type: [string, 'null'], format: date-time}
...
    ApprovalRequester:
      type: object
      required: [id, email]
      description: Requester identity only; not a workspace User and has no status field.
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        display_name: {type: [string, 'null']}
    ApprovalRequestedRole:
      type: object
      required: [id, code, display_name]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        display_name: {type: string}
    ApprovalRequest:
      type: object
      required: [id, request_type, target_id, status, requester, created_at]
      properties:
        id: {type: string, format: uuid}
        request_type: {type: string, enum: [resource_access, plugin_installation]}
        target_id: {type: string}
        status: {type: string, enum: [pending, approved, denied, cancelled]}
        requester: {$ref: '#/components/schemas/ApprovalRequester'}
        requested_role: {anyOf: [{$ref: '#/components/schemas/ApprovalRequestedRole'}, {type: 'null'}]}
        details: {type: object, additionalProperties: true}
        reason: {type: [string, 'null']}
        decision_note: {type: [string, 'null']}
        decided_by_user_id: {type: [string, 'null'], format: uuid}
        decided_at: {type: [string, 'null'], format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: [string, 'null'], format: date-time}
    ApprovalRequestCreate:
      type: object
      required: [request_type, target_id]
      properties:
        request_type: {type: string, enum: [resource_access, plugin_installation]}
        target_id: {type: string}
        details: {type: object, additionalProperties: true}
        reason: {type: [string, 'null'], maxLength: 4000}
    ApprovalRequestUpdate:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [approved, denied, cancelled]}
        decision_note: {type: [string, 'null'], maxLength: 4000}
    ApprovalRequestPage:
      allOf:
        - {$ref: '#/components/schemas/PageFields'}
        - type: object
          required: [items]
          properties: {items: {type: array, items: {$ref: '#/components/schemas/ApprovalRequest'}}}
...
    AccessSessionRecord:
      type: object
      description: Session metadata only; never tokens, token_version, metadata, or provider subjects.
      required: [id, user, authentication_method, entry, status, started_at, last_seen_at, ended_at, end_reason, expires_at, current]
      properties:
        id: {type: string, format: uuid}
        user:
          type: object
          required: [id, email, display_name]
          properties:
            id: {type: string, format: uuid}
            email: {type: [string, 'null']}
            display_name: {type: [string, 'null']}
        authentication_method: {type: string}
        entry: {type: string, enum: [sign_in, workspace_switch]}
        status:
          type: string
          enum: [active, expired, revoked, superseded]
          description: Effective status; a stored-active row past its absolute or idle expiry is expired.
        started_at: {type: string, format: date-time}
        last_seen_at: {type: [string, 'null'], format: date-time}
        ended_at:
          type: [string, 'null']
          format: date-time
          description: For an effectively expired row still stored active, the expiry that lapsed.
        end_reason: {type: [string, 'null']}
        expires_at: {type: string, format: date-time}
        current: {type: boolean, description: The caller's own session}
    AccessSessionPage:
      allOf:
        - {$ref: '#/components/schemas/PageFields'}
        - type: object
          required: [items]
          properties: {items: {type: array, items: {$ref: '#/components/schemas/AccessSessionRecord'}}}
...